requirements.txt
numpy < 1.19
scipy >= 1.5
numba == 0.48
tensorflow < 2.0 
pycuda
cython-bbox
asyncio-nats-client
PyTurboJPEG
protobuf